Fix intermittent missed movement keys
Issue was that multiple calls to update occurred between calls to fixed update and if the input on the first update was registered, it would be cleared (variable set to false) on the next call to update (before fixed update) resulting in that key press being lost. Hence, checking only allowing the boolean variables to change from false to true in update, and true to false in fixed update.