Refactor and improve reliability of movement input
Set the next movement direction based on the input key, then execute this (provided it does not take the player out of bounds) in FixedUpdate. Now it is not possible for the player to be requesting to move in multiple directions at once (e.g. moveRight and moveUp both true), since there a vector is set instead of 4 boolean variables.